pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nuget-syncer"
version = "0.1.0"
description = "Sync NuGet packages from Azure DevOps to GitHub Packages"
authors = [
    {name = "Keyfactor", email = "support@keyfactor.com"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.28"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.12.0"
]

[tool.setuptools]
packages = ["scripts"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100
